test(e2e): lift feature-flags mock to global Playwright fixture
Fixes #146
