+++ e2e/cross-page-profile-tools.spec.ts
@@ -1,4 +1,5 @@
-import { test, expect, Page } from '@playwright/test'
+import { test, expect } from './fixtures/base'
+import type { Page } from '@playwright/test'
 import {
   mutateProfile,
   seedBudgetCsvsForYear,
