+++ e2e/net-worth.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '@playwright/test'
+import { test, expect } from './fixtures/base'
 import { NetWorthPage } from './pages/net-worth.page'
 import {
   ACCOUNTS,
